vcs: fix `-go-internal-cd` (#10)
The command directory was always being overridden by the context
directory. This actually meant that some VCS commands were running
in the wrong place (e.g. doing a submodule update on the CWD rather
than the cloned repo).

Often, this wasn't noticeable because the CWD was _also_ a Git repo;
in fact, that's why tests didn't catch this since the `go-get` dir
is a Git repo. Tests now set the CWD to an empty tmpdir to avoid
this.